I tried fixing seminar 4
I'm very tired of this, but after the last few days of going back and forwards on what and how to do stuff, I just decided to put logic in the controller. I now just need to fix so that the terminate commands cant terminate non-existing instruments. We'll need a new sql statement for this though.